CUDA: parallelizing kick(), naively
We use the most straightforward way to parallelize kick().  We time
the performance for different version of the code.  The GPU
performance fluctuates a lot...

            with -O0   with -O3       CUDA
      n      ns/step    ns/step    ns/step
  ----------------------------------------
      4          300        120       7879
      8         1176        543       9587
     16         4590       2253      12818
     32        19390       8955      17655
     64        76569      36530      30073
    128       329548     157217      46547
    256      1299920     618476      83311
    512      5337840    2487450     159845

We can only see the performance scale only linearly with n.  For
n > 128, the GPU code runs only a few times faster than the (compiler
optimized) CPU code.  We need to dig deeper into the GPU architecture
to gain more performance.